on high loads (100% disk usage + > 80% CPU + >96 % RAM used) the errors started flooding my screen. Turns out that if drive is not ready, the screenshot is not written, but the trigger is being called..
So to fix this, I changed the bool to lowest-memory usage option possible - int (well could have also byte, but nowdays its 64-bit computing anyways..)
Besides IMHO it is more straightforward than what I offered yesterday - yesterdays variant had the unwelcome side effect of making the SaveScreenshot aware of some additional object property (the Optimization).

Now it is closer to single responsibility pattern - all it knows is whether the screenshot taking had an errorlevel or not (which now can easily be turned to a trigger call at exceptional cases)

For success it returns 0. To make it searchable I put it under pattern return returnFlag (opposed to return 2 or return 32, which is hard to overlook and slower to grasp than enums.)

For editors no code change just updated the xml comment to match the new method signature.